requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.2
django-js-asset==2.2.0
django-unfold==0.33.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1
whitenoise==6.7.0
